move filters class to own file
easier to see what is going on
